Remove unnecessary add_bias and remove_bias.
Add a input boolean arg in vercat called "reserve". vercat reserve one
more row if it's true.

Add a input variable for versplit so that it can safely ignore the
additional row in the input data.

Potential crash. What if user choose a conv but not a sub-sampling
following it ??? FIXME